+++ composer.json
@@ -10,12 +10,29 @@
         }
     ],
     "autoload": {
+        "classmap": [
+            "src/"
+        ]
+    },
+    "autoload-dev": {
         "psr-4": {
-            "WooCommerce\\Migrator\\": "src/"
+            "WooCommerce\\Migrator\\Tests\\": "tests/"
         }
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
-        "yoast/phpunit-polyfills": "^1.0"
+        "yoast/phpunit-polyfills": "^1.0",
+        "woocommerce/woocommerce-sniffs": "^1.0"
+    },
+    "config": {
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
+    },
+    "scripts": {
+        "lint": "phpcs",
+        "format": "phpcbf",
+        "test:setup": "bash bin/install-wp-tests.sh ${DB_NAME:-test_db} ${DB_USER:-root} ${DB_PASS:-''} ${DB_HOST:-localhost}",
+        "test": "phpunit"
     }
 }
