+++ phpcs.xml.dist
@@ -1,0 +1,35 @@
+<?xml version="1.0"?>
+<ruleset name="WooCommerce Migrator">
+	<description>WooCommerce Migrator coding standards.</description>
+
+	<file>./src</file>
+	<file>./tests</file>
+	<file>./woocommerce-migrator.php</file>
+
+	<!-- Exclude vendor directory -->
+	<exclude-pattern>/vendor/*</exclude-pattern>
+
+	<arg name="extensions" value="php"/>
+	<arg name="colors"/>
+	<arg value="s"/>
+
+	<!-- Set the PHP test version for PHPCompatibility checks. -->
+	<config name="testVersion" value="7.0-"/>
+
+	<!-- Rules -->
+	<rule ref="WooCommerce-Core">
+		<!-- We may want to relax this in the future, but for now, let's keep it. -->
+		<exclude name="WordPress.Files.FileName"/>
+	</rule>
+
+	<rule ref="WordPress.WP.I18n">
+		<properties>
+			<property name="text_domain" type="array" value="woocommerce-migrator" />
+		</properties>
+	</rule>
+
+	<!-- Run the PHPCompatibility check. -->
+	<rule ref="PHPCompatibilityWP">
+		<exclude name="PHPCompatibility.FunctionDeclarations.NewParamTypeDeclarations.objectFound"/>
+	</rule>
+</ruleset>
