+++ tests/TestCase.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Base test case
+ *
+ * @package WooCommerce\Migrator\Tests
+ */
+
+declare( strict_types=1 );
+
+namespace WooCommerce\Migrator\Tests;
+
+/**
+ * Our main test case class.
+ */
+abstract class TestCase extends \PHPUnit\Framework\TestCase {
+
+}
